+++ internal/e2e/doc.go
@@ -1,0 +1,2 @@
+// package e2e contains end to end tests
+package e2e
