+++ internal/e2e/testdata/Repository_metadata_Write_rbac.txt
