+++ internal/e2e/testdata/bufio_buffer.txt
