+++ internal/e2e/testdata/bytes_buffer.txt
