+++ internal/e2e/testdata/graphql_type_User.txt
